backend/pyproject.toml
[project]
name = "aicompass-backend"
version = "0.1.0"
description = "FastAPI backend for AI Compass"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pymongo==4.6.1",
    "pydantic>=2.10.0",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.1.2",
    "python-multipart==0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
dev = "app.dev:main"
prod = "app.prod:main"
